styling of messages and navbar
