+++ minitwit-frontend/src/components/Timeline/Timeline.css
@@ -1,13 +1,5 @@
-.timeline-container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    background-color: #282c34;
-    padding: 20px 0 100px;
-}
-
 .timeline {
-    width: 60%;
+    width: 100%;
     min-width: 270px;
 }
