+++ minitwit-frontend/src/pages/Timeline/TimelinePage.css
@@ -4,4 +4,5 @@
     align-items: center;
     background-color: #282c34;
     padding: 20px 0 100px;
+    width: 50%;
 }
